Move tabs configuration
