--- apps/meteor/app/mentions-flextab/client/index.ts
@@ -1,2 +1,0 @@
-import './actionButton';
-import './tabBar';
