--- apps/meteor/app/message-pin/client/index.ts
@@ -1,3 +1,0 @@
-import './actionButton';
-import './pinMessage';
-import './tabBar';
