--- apps/meteor/app/message-star/client/index.ts
@@ -1,3 +1,0 @@
-import './actionButton';
-import './starMessage';
-import './tabBar';
