+++ apps/meteor/client/importPackages.ts
@@ -27,12 +27,9 @@ import '../app/lib/client';
 import '../app/livestream/client';
 import '../app/logger/client';
 import '../app/markdown/client';
-import '../app/mentions-flextab/client';
 import '../app/message-attachments/client';
 import '../app/message-mark-as-unread/client';
-import '../app/message-pin/client';
 import '../app/message-snippet/client';
-import '../app/message-star/client';
 import '../app/nextcloud/client';
 import '../app/oauth2-server-config/client';
 import '../app/oembed/client';
