+++ apps/meteor/client/main.ts
@@ -10,11 +10,7 @@ import './importPackages';
 
 import '../ee/client';
 import './templateHelpers';
-import './methods/hideRoom';
-import './methods/openRoom';
-import './methods/setUserActiveStatus';
-import './methods/toggleFavorite';
-import './methods/updateMessage';
+import './methods';
 import './startup';
 import './views/admin';
 import './views/account';
