+++ apps/meteor/client/methods/index.ts
@@ -1,0 +1,8 @@
+import './hideRoom';
+import './openRoom';
+import './pinMessage';
+import './setUserActiveStatus';
+import './starMessage';
+import './toggleFavorite';
+import './unpinMessage';
+import './updateMessage';
