+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -1,0 +1,10 @@
+import './jumpToMessage';
+import './jumpToPinMessage';
+import './jumpToStarMessage';
+import './permalinkPinned';
+import './permalinkStar';
+import './pinMessage';
+import './readReceipt';
+import './starMessage';
+import './unpinMessage';
+import './unstarMessage';
