+++ apps/meteor/client/startup/actionButtons/readReceipt.ts
@@ -1,11 +1,11 @@
 import { Meteor } from 'meteor/meteor';
 import { Tracker } from 'meteor/tracker';
 
-import { settings } from '../../app/settings/client';
-import { MessageAction } from '../../app/ui-utils/client';
-import { imperativeModal } from '../lib/imperativeModal';
-import { messageArgs } from '../lib/utils/messageArgs';
-import ReadReceiptsModal from '../views/room/modals/ReadReceiptsModal';
+import { settings } from '../../../app/settings/client';
+import { MessageAction } from '../../../app/ui-utils/client';
+import { imperativeModal } from '../../lib/imperativeModal';
+import { messageArgs } from '../../lib/utils/messageArgs';
+import ReadReceiptsModal from '../../views/room/modals/ReadReceiptsModal';
 
 Meteor.startup(() => {
 	Tracker.autorun(() => {
