+++ apps/meteor/client/startup/contextualBar/index.ts
@@ -1,1 +1,4 @@
-import './exportMessages';
+import './exportMessagesTab';
+import './mentionsTab';
+import './pinnedMessagesTab';
+import './starredMessagesTab';
