+++ apps/meteor/client/startup/contextualBar/mentionsTab.ts
@@ -1,12 +1,12 @@
 import { lazy } from 'react';
 
-import { addAction } from '../../../client/views/room/lib/Toolbox';
+import { addAction } from '../../views/room/lib/Toolbox';
 
 addAction('mentions', {
 	groups: ['channel', 'group', 'team'],
 	id: 'mentions',
 	title: 'Mentions',
 	icon: 'at',
-	template: lazy(() => import('../../../client/views/room/contextualBar/Mentions')),
+	template: lazy(() => import('../../views/room/contextualBar/MentionsTab')),
 	order: 9,
 });
