+++ apps/meteor/client/startup/contextualBar/pinnedMessagesTab.ts
@@ -1,8 +1,8 @@
-import { lazy, useMemo } from 'react';
-import { useSetting } from '@rocket.chat/ui-contexts';
 import { isRoomFederated } from '@rocket.chat/core-typings';
+import { useSetting } from '@rocket.chat/ui-contexts';
+import { lazy, useMemo } from 'react';
 
-import { addAction } from '../../../client/views/room/lib/Toolbox';
+import { addAction } from '../../views/room/lib/Toolbox';
 
 addAction('pinned-messages', ({ room }) => {
 	const pinningAllowed = useSetting('Message_AllowPinning');
@@ -15,7 +15,7 @@ addAction('pinned-messages', ({ room }) => {
 						id: 'pinned-messages',
 						title: 'Pinned_Messages',
 						icon: 'pin',
-						template: lazy(() => import('../../../client/views/room/contextualBar/PinnedMessages')),
+						template: lazy(() => import('../../views/room/contextualBar/PinnedMessagesTab')),
 						...(federated && {
 							'data-tooltip': 'Pinned_messages_unavailable_for_federation',
 							'disabled': true,
