+++ apps/meteor/client/startup/contextualBar/starredMessagesTab.ts
@@ -1,12 +1,12 @@
 import { lazy } from 'react';
 
-import { addAction } from '../../../client/views/room/lib/Toolbox';
+import { addAction } from '../../views/room/lib/Toolbox';
 
 addAction('starred-messages', {
 	groups: ['channel', 'group', 'direct', 'direct_multiple', 'team'],
 	id: 'starred-messages',
 	title: 'Starred_Messages',
 	icon: 'star',
-	template: lazy(() => import('../../../client/views/room/contextualBar/StarredMessages')),
+	template: lazy(() => import('../../views/room/contextualBar/StarredMessagesTab')),
 	order: 10,
 });
