+++ apps/meteor/client/startup/index.ts
@@ -1,4 +1,6 @@
+import '../lib/rooms/roomTypes';
 import './absoluteUrl';
+import './actionButtons';
 import './afterLogoutCleanUp';
 import './appRoot';
 import './banners';
@@ -22,7 +24,6 @@ import './oauth';
 import './openedRoom';
 import './otr';
 import './readMessage';
-import './readReceipt';
 import './reloadRoomAfterLogin';
 import './renderMessage';
 import './renderNotification';
@@ -38,4 +39,3 @@ import './UserDeleted';
 import './userRoles';
 import './usersObserve';
 import './userStatusManuallySet';
-import '../lib/rooms/roomTypes';
