--- apps/meteor/client/views/room/contextualBar/Mentions/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Mentions';
