--- apps/meteor/client/views/room/contextualBar/PinnedMessages/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PinnedMessages';
