--- apps/meteor/client/views/room/contextualBar/StarredMessages/index.ts
@@ -1,1 +1,0 @@
-export { default } from './StarredMessages';
